+++ src/format/sinks/SinkCFILE.cpp
@@ -2,9 +2,9 @@
 
 namespace format
 {
-size_t cfile_sink::Write(Context&, const Context::SlicesT& slices, int flags)
+int cfile_sink::Write(Context&, const Context::SlicesT& slices, int flags)
 {
-	size_t size = 0;
+	int size = 0;
 	for(auto& s : slices) {
 		if(fwrite(s.data, s.size, 1, m_File) != 1)
 			return size;
